feat: refactor playload type
