+++ api/src/extensions/channels/web/wrapper.ts
@@ -9,9 +9,9 @@
 import { Attachment } from '@/attachment/schemas/attachment.schema';
 import EventWrapper from '@/channel/lib/EventWrapper';
 import { ChannelName } from '@/channel/types';
+import { PayloadType } from '@/chat/schemas/types/button';
 import {
   IncomingMessageType,
-  PayloadType,
   StdEventType,
   StdIncomingMessage,
 } from '@/chat/schemas/types/message';
